Implement a message content size adjustment
The content display control of the KryptonTaskDialog is changed
based on the size of the content.  If the content contains more
than 20 lines of text, the control changes to a multi-line text
container.  The size of the KryptonTaskDialog is also limitted
to 0.6 size of the display area to fix the issue of displaying
very large dialog boxes that don't fit in a screen.